refactor: split generator facade into focused services
Extract MDX reconciliation coordination and pure OpenAPI rendering while preserving the public facade and transactional rollback behavior. Split the regression suites by behavior to keep coverage focused.
